+++ src/Pusher/PusherFactory.h
@@ -15,15 +15,13 @@
 #include "PusherPonderomotivePositionBoris.h"
 #include "PusherVay.h"
 #include "PusherBorisNR.h"
-#include "PusherRRLL.h"
 #include "PusherHigueraCary.h"
 #include "PusherPhoton.h"
 
-#ifdef _VECTO
-#include "PusherBorisV.h"
-#include "PusherPonderomotiveBorisV.h"
-#include "PusherPonderomotivePositionBorisV.h"
-#endif
+// #ifdef _VECTO
+// #include "PusherPonderomotiveBorisV.h"
+// #include "PusherPonderomotivePositionBorisV.h"
+// #endif
 
 #include "Params.h"
 #include "Species.h"
@@ -50,15 +48,16 @@ public:
         // Particle of matter
         if( species->mass_ > 0 ) {
             // assign the correct Pusher to Push
+            // Pusher of Boris
             if( species->pusher_name_ == "boris" ) {
-                if( !species->vectorized_operators && !params.cell_sorting ) {
+                // if( !species->vectorized_operators && !params.cell_sorting ) {
                     Push = new PusherBoris( params, species );
-                }
-#ifdef _VECTO
-                else {
-                    Push = new PusherBorisV( params, species );
-                }
-#endif
+                // }
+// #ifdef _VECTO
+//                 else {
+//                     Push = new PusherBorisV( params, species );
+//                 }
+// #endif
             } else if( species->pusher_name_ == "ponderomotive_boris" ) {
             
                 int n_envlaser = params.Laser_Envelope_model;
@@ -70,23 +69,22 @@ public:
                     ERROR( "if ponderomotive_boris pusher is chosen for a species, the flag ponderomotive_dynamics for that species must be set to true." );
                 }
                 
-                if( !species->vectorized_operators && !params.cell_sorting ) {
+                // if( !species->vectorized_operators && !params.cell_sorting ) {
                     Push = new PusherPonderomotiveBoris( params, species );
-                }
-#ifdef _VECTO
-                else {
-                    Push = new PusherPonderomotiveBorisV( params, species );
-                }
-#endif
+//                 }
+// #ifdef _VECTO
+//                 else {
+//                     Push = new PusherPonderomotiveBorisV( params, species );
+//                 }
+// #endif
+            // Non-relativistic Boris pusher
             } else if( species->pusher_name_ == "borisnr" ) {
                 Push = new PusherBorisNR( params, species );
             }
-            /*else if ( species->pusher_name_ == "rrll" )
-            {
-                Push = new PusherRRLL( params, species );
-            }*/
+            // Pusher of J.L. Vay
             else if( species->pusher_name_ == "vay" ) {
                 Push = new PusherVay( params, species );
+            // Pusher of Higuera Cary
             } else if( species->pusher_name_ == "higueracary" ) {
                 Push = new PusherHigueraCary( params, species );
             } else {
@@ -122,14 +120,14 @@ public:
         if( species->mass_ > 0 ) {
             // assign the correct Pusher to Push_ponderomotive_position
             if( species->pusher_name_ == "ponderomotive_boris" ) {
-                if( !species->vectorized_operators && !params.cell_sorting ) {
+                // if( !species->vectorized_operators && !params.cell_sorting ) {
                     Push_ponderomotive_position = new PusherPonderomotivePositionBoris( params, species );
-                }
-#ifdef _VECTO
-                else {
-                    Push_ponderomotive_position = new PusherPonderomotivePositionBorisV( params, species );
-                }
-#endif
+//                 }
+// #ifdef _VECTO
+//                 else {
+//                     Push_ponderomotive_position = new PusherPonderomotivePositionBorisV( params, species );
+//                 }
+// #endif
             }
             
             else {